feat: update word types to match the needs of the word pantry UI
Updated the partitions to output display data that is lighter than the full parsed data and provides only what the word pantry UI needs.

Added a spellings to wordId for autocomplete.

Added a flag to the wordDisplayData type so we can know whether to make the references call at all -- if there are no references, we can skip that call.

Signed-off-by: Kori Ryter <kryter@gmail.com>